requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.2
django-heroku==0.3.1
gunicorn==23.0.0
narwhals==1.41.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
plotly==6.1.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
whitenoise==6.9.0
